fix: type related compiling error
This commit fixes a compiling error related to typing.

+++ tablec.c
@@ -156,7 +156,7 @@ void tablec_del(struct tablec_ht *tablec, char *key) {
 }
 
 int tablec_full(struct tablec_ht *tablec) {
-  return tablec->capacity == tablec->length ? -1 : tablec->capacity - tablec->length;
+  return (int)(tablec->capacity == tablec->length) ? -1 : (int)(tablec->capacity - tablec->length);
 }
 
 void tablec_cleanup(struct tablec_ht *tablec) {
